requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
pydantic==2.12.0
pydantic-settings==2.6.1
pydantic[email]==2.12.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
asyncpg==0.30.0
psycopg[binary]==3.2.3
fastapi-cache2[redis]
celery
flower
sqladmin
itsdangerous
